requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.1
attrs==21.2.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
dateparser==1.1.0
frozenlist==1.2.0
idna==3.3
iniconfig==1.1.1
multidict==5.2.0
numpy==1.21.4
packaging==21.2
pandas==1.3.4
pep8==1.7.1
pluggy==1.0.0
py==1.11.0
pybit==1.3.2
pycodestyle==2.8.0
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.5
pytest-asyncio==0.16.0
python-binance==1.0.15
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.10
requests==2.26.0
six==1.16.0
toml==0.10.2
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
ujson==4.2.0
urllib3==1.26.7
websocket-client==1.2.1
websockets==9.1
yarl==1.7.2
